refactor: extract reusable audited-output primitives
